+++ src/greenlet/greenlet_cpython_compat.hpp
@@ -9,33 +9,6 @@
 #define PY_SSIZE_T_CLEAN
 #include "Python.h"
 
-// These enable writing template functions or classes specialized
-// based on the Python version. Write both versions of the function,
-// one with the WHEN version, one with the WHEN_NOT version.
-// Instantiate the template using the G_IS_PY37 macro.
-struct GREENLET_WHEN_PY37
-{
-    typedef GREENLET_WHEN_PY37* Yes;
-    // We really just want an alias, `using Yes = IsIt`,
-    // but old MSVC for Py27 doesn't support that.
-    typedef GREENLET_WHEN_PY37* IsIt;
-};
-
-struct GREENLET_WHEN_NOT_PY37
-{
-    typedef GREENLET_WHEN_NOT_PY37* No;
-    typedef GREENLET_WHEN_NOT_PY37* IsIt;
-};
-
-
-#if PY_VERSION_HEX >= 0x030700A3
-#    define GREENLET_PY37 1
-typedef GREENLET_WHEN_PY37 G_IS_PY37;
-#else
-#    define GREENLET_PY37 0
-typedef GREENLET_WHEN_NOT_PY37 G_IS_PY37;
-#endif
-
 
 #if PY_VERSION_HEX >= 0x30A00B1
 /*
@@ -48,11 +21,7 @@ We have to save and restore this as well.
 #    define GREENLET_USE_CFRAME 0
 #endif
 
-#if PY_VERSION_HEX >= 0x30C0000
-#    define GREENLET_PY312 1
-#else
-#    define GREENLET_PY312 0
-#endif
+
 
 #if PY_VERSION_HEX >= 0x30B00A4
 /*
@@ -70,6 +39,13 @@ https://bugs.python.org/issue46090). Summary of breaking internal changes:
 #    define GREENLET_PY311 1
 #else
 #    define GREENLET_PY311 0
+#endif
+
+
+#if PY_VERSION_HEX >= 0x30C0000
+#    define GREENLET_PY312 1
+#else
+#    define GREENLET_PY312 0
 #endif
 
 #ifndef Py_SET_REFCNT
@@ -98,47 +74,19 @@ https://bugs.python.org/issue39573 */
 #ifndef Py_TPFLAGS_HAVE_NEWBUFFER
   #define Py_TPFLAGS_HAVE_NEWBUFFER 0
 #endif
-#ifndef Py_TPFLAGS_HAVE_FINALIZE
-  #define Py_TPFLAGS_HAVE_FINALIZE 0
-#endif
+
 #ifndef Py_TPFLAGS_HAVE_VERSION_TAG
    #define Py_TPFLAGS_HAVE_VERSION_TAG 0
 #endif
 
 #define G_TPFLAGS_DEFAULT Py_TPFLAGS_DEFAULT | Py_TPFLAGS_HAVE_VERSION_TAG | Py_TPFLAGS_CHECKTYPES | Py_TPFLAGS_HAVE_NEWBUFFER | Py_TPFLAGS_HAVE_GC
 
-#if PY_MAJOR_VERSION >= 3
-#    define GNative_FromFormat PyUnicode_FromFormat
-#else
-#    define GNative_FromFormat PyString_FromFormat
-#endif
-
-#if PY_MAJOR_VERSION >= 3
-#    define Greenlet_Intern PyUnicode_InternFromString
-#else
-#    define Greenlet_Intern PyString_InternFromString
-#endif
 
 #if PY_VERSION_HEX < 0x03090000
 // The official version only became available in 3.9
 #    define PyObject_GC_IsTracked(o) _PyObject_GC_IS_TRACKED(o)
 #endif
 
-#if PY_MAJOR_VERSION < 3
-struct PyModuleDef {
-    int unused;
-    const char* const m_name;
-    const char* m_doc;
-    Py_ssize_t m_size;
-    PyMethodDef* m_methods;
-    // Then several more fields we're not currently using.
-};
-#define PyModuleDef_HEAD_INIT 1
-PyObject* PyModule_Create(PyModuleDef* m)
-{
-    return Py_InitModule(m->m_name, m->m_methods);
-}
-#endif
 
 // bpo-43760 added PyThreadState_EnterTracing() to Python 3.11.0a2
 #if PY_VERSION_HEX < 0x030B00A2 && !defined(PYPY_VERSION)