+++ src/greenlet/tests/_test_extension.c
@@ -203,8 +203,8 @@ static PyMethodDef test_methods[] = {
     {NULL, NULL, 0, NULL}
 };
 
-#if PY_MAJOR_VERSION >= 3
-#    define INITERROR return NULL
+
+#define INITERROR return NULL
 
 static struct PyModuleDef moduledef = {PyModuleDef_HEAD_INIT,
                                        TEST_MODULE_NAME,
@@ -218,27 +218,14 @@ static struct PyModuleDef moduledef = {PyModuleDef_HEAD_INIT,
 
 PyMODINIT_FUNC
 PyInit__test_extension(void)
-#else
-#    define INITERROR return
-PyMODINIT_FUNC
-init_test_extension(void)
-#endif
 {
     PyObject* module = NULL;
-
-#if PY_MAJOR_VERSION >= 3
     module = PyModule_Create(&moduledef);
-#else
-    module = Py_InitModule(TEST_MODULE_NAME, test_methods);
-#endif
 
     if (module == NULL) {
-        INITERROR;
+        return NULL;
     }
 
     PyGreenlet_Import();
-
-#if PY_MAJOR_VERSION >= 3
     return module;
-#endif
 }