+++ src/greenlet/tests/test_greenlet.py
@@ -578,8 +578,6 @@ class TestGreenlet(TestCase):
         self.assertEqual(value, 42)
 
     def test_tuple_subclass(self):
-        # XXX: This is failing on Python 2 with a SystemError: error return without exception set
-
         # The point of this test is to see what happens when a custom
         # tuple subclass is used as an object passed directly to the C
         # function ``green_switch``; part of ``green_switch`` checks
@@ -591,13 +589,10 @@ class TestGreenlet(TestCase):
         # `apply` function directly passes the given args tuple object
         # to the underlying function, whereas the Python 3 version
         # unpacks and repacks into an actual tuple. This could still
-        # happen using the C API on Python 3 though.
-        if sys.version_info[0] > 2:
-            # There's no apply in Python 3.x
-            def _apply(func, a, k):
-                func(*a, **k)
-        else:
-            _apply = apply # pylint:disable=undefined-variable
+        # happen using the C API on Python 3 though. We should write a
+        # builtin version of apply() ourself.
+        def _apply(func, a, k):
+            func(*a, **k)
 
         class mytuple(tuple):
             def __len__(self):